+++ src/utils/messaging.ts
@@ -12,7 +12,9 @@ export const showErrorMessage = (
   message: string,
   withButton: boolean = true
 ): void => {
-  if (!settings.showErrorNotifications) return;
+  if (!settings.showErrorNotifications) {
+    return;
+  };
   if (!withButton) {
     vscode.window.showErrorMessage(message);
     return;
@@ -45,7 +47,7 @@ export const showReloadMessage = (
   message: string,
   type: "default" | "initial" = "default"
 ): void => {
-  if (!settings.showReloadNotification) return;
+  if (!settings.showReloadNotification) {return;}
 
   if (type === "default") {
     vscode.window.showInformationMessage(message);